Implementation of the Dialog-Button-Handler
The name of the POI is set in the dialog title and
in the EditText field of the dialog.

By clicking on delete, the POI is removed form
the database and the POIContainer. After this the POIList View
is refreshed.

Renaming the POI-Name affects the datatabse and the POIContainer, as
well as the UI.

Use the uiRefresh method form the interface IUIRefreshable to refresh
the UI.